+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 4.3.3
+- Add the ability to disable Features added to Vanilla Biomes on NeoForge
 
 # 4.3.2
 - Fix ManOWar Crash when using Geckolib 5.4.4